--- src/components/RepositoryAutocompleteContainer.tsx
@@ -1,48 +1,0 @@
-import React from "react";
-import { Box, Text } from "ink";
-import Spinner from "ink-spinner";
-import RepositoryAutocomplete from "./RepositoryAutocomplete.js";
-import { useRepositories } from "../hooks/useRepositories.js";
-import { Repository } from "../lib/clients/baz.js";
-
-interface RepositoryAutocompleteContainerProps {
-  onSelect: (repo: Repository) => void;
-  initialRepoId?: string;
-}
-
-const RepositoryAutocompleteContainer: React.FC<
-  RepositoryAutocompleteContainerProps
-> = ({ onSelect, initialRepoId }) => {
-  const { data, loading, error } = useRepositories();
-
-  if (loading) {
-    return (
-      <Box>
-        <Text color="blue">
-          <Spinner type="dots" />
-        </Text>
-        <Text color="blue"> Fetching repositories...</Text>
-      </Box>
-    );
-  }
-
-  if (error) {
-    return (
-      <Box flexDirection="column">
-        <Text color="red" bold>
-          ❌ Error: {error}
-        </Text>
-      </Box>
-    );
-  }
-
-  return (
-    <RepositoryAutocomplete
-      repositories={data}
-      onSelect={onSelect}
-      initialRepoId={initialRepoId}
-    />
-  );
-};
-
-export default RepositoryAutocompleteContainer;
